back/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
PyMuPDF==1.22.5
openai==1.39.0
python-multipart==0.0.8
python-dotenv==0.21.1
